mavlink_bindgen

Function generate

pub fn generate<P1: AsRef<Path>, P2: AsRef<Path>>(
    xml_definitions: XmlDefinitions<P1>,
    destination_dir: P2,
) -> Result<GeneratedBindings, BindGenError>

Generate Rust MAVLink dialect binding for dialects present in the given xml_definitions into destination_dir.

If successful returns paths of generated bindings linked to their dialect definitions files.