mavlink_bindgen

Module parser

Structs

MavEnum

MavEnumEntry

MavField

MavMessage

MavProfile

Enums

MavType

MavXmlElement

Functions

extra_crc

CRC operates over names of the message and names of its fields Hence we have to preserve the original uppercase names delimited with an underscore For field names, we replace “type” with “mavtype” to make it rust compatible (this is needed for generating sensible rust code), but for calculating crc function we have to use the original name “type”

generate

Generate protobuf represenation of mavlink message set Generate rust representation of mavlink message set with appropriate conversion methods

parse_profile