mavlink_bindgen::parser

Function extra_crc

pub fn extra_crc(msg: &MavMessage) -> u8

CRC operates over names of the message and names of its fields Hence we have to preserve the original uppercase names delimited with an underscore For field names, we replace “type” with “mavtype” to make it rust compatible (this is needed for generating sensible rust code), but for calculating crc function we have to use the original name “type”